functions/dummy_function/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-core==2.* 